+++ src/react-components/src/components/01-atoms/Images/Image/Image.stories.tsx
@@ -12,3 +12,6 @@ export const twoByOne = () => [<Image src="https://placeimg.com/200/100/arch" is
 <Image src="https://placeimg.com/400/200/arch" isDecorative={true}/>,
 <Image src="https://placeimg.com/1600/800/arch" isDecorative={true}/>,
 <Image src="https://placeimg.com/2000/1000/arch" isDecorative={true}/> ];
+
+export const threeByFour = () => [<Image src="https://placeimg.com/150/200/arch" isDecorative={true}/>,
+<Image src="https://placeimg.com/300/400/arch" isDecorative={true}/> ];